+++ Runtime/Tweens/Components/TweenState.cs
@@ -6,7 +6,7 @@ namespace Timespawn.EntityTween.Tweens
 {
     public struct TweenState : IBufferElementData, ITweenId
     {
-        public const byte LOOP_COUNT_INFINITE = 0;
+        internal const byte LOOP_COUNT_INFINITE = 0;
         private const byte LOOP_COUNT_PENDING_DESTROY = byte.MaxValue;
 
         public int Id;